requirements.txt
torch>=1.9.0
torchvision>=0.10.0
captum>=0.5.0
lime>=0.2.0.1
shap>=0.40.0
matplotlib>=3.5.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
jupyter>=1.0.0
pillow>=8.0.0
scikit-learn>=1.0.0
requests>=2.25.0
tqdm>=4.62.0
gradio>=3.0.0
